+++ frontend/public/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 
+- Fixed issue where Plex WebSocket client would not properly stop when changed from the Settings Page.
 - Fixed issue where migration v1 to v2 was taking too long.
 - Fixed PANIC error when bad request was returned by Plex.